\name{asht}
\alias{asht}
\title{
  Internal function for simultaneous asymptotic intervals
}
\description{
  Internal function for simultaneous asymptotic intervals
}
\references{
  Fritsch, K. S., Hsu, J. C. (1999) Multiple comparison of
    entropies with application to dinosaur biodiversity. \emph{Biometrics},
  55, 4, 1300--1305.

  Rogers, J. A., Hsu, J. C. (2001) Multiple comparisons of
    biodiversity. \emph{Biometrical Journal}, 43, 5, 617--625.
}
\note{
  Only internal function. Use function \code{\link{sbdiv}} instead
}
\keyword{ htest }

