\name{estThetaRow}
\alias{estThetaRow}
\title{
  Internal function
}
\description{
  Internal function for method \code{\link{WYht}} in function
  \code{\link{sbdiv}}. Calculates the specified diversity index for every
  replicated sample in each treatment group.
}
\usage{
estThetaRow(X, f, theta)
}
\arguments{
  \item{X}{
    Matrix with dimension \eqn{n \times p}.
  }
  \item{f}{
    Factorial variable containing treatment groups.
}
  \item{theta}{
}
}
\keyword{ misc }
