\name{rpht}
\alias{rpht}
\title{
  Internal function for simultaneous bayesian bootstrap intervals
}
\description{
  Internal function for simultaneous bayesian bootstrap intervals
}
\references{
  Besag, J., Green, P. J., Higdon, D., Mengersen, K. (1995)
  Bayesian computation and stochastic systems (with discussion)
  . \emph{Statistical Science}, 10, 3--66.
}
\note{
  Only internal function. Use function \code{\link{sbdiv}} instead
}
\keyword{ htest }
