\name{Simpson}
\alias{Simpson}
\title{
  Internal function for Simpson estimator
}
\description{
  Calculates Simpson's index on probability vector \eqn{p}
}
\usage{
Simpson(p)
}
\arguments{
  \item{p}{
    Probability vector \eqn{x_s/n}
}
}
\value{
  Simpson's index
}
\note{
  Only for internal use
}
\keyword{ misc }
