\name{tsht}
\alias{tsht}
\title{
  Internal function for simultaenous bootstrap intervals
}
\description{
  Internal function for simultaenous bootstrap intervals based on summed
  up counts for every species.
}
\references{
  Beran, R. (1988) Balanced simultaneous confidence
    sets. \emph{Journal of the American Statistical Association}, 83, 679--686.
}
\note{
  Only internal function. Use function \code{\link{sbdiv}} instead
}
\keyword{ htest }
