\name{WYht}
\alias{WYht}
\title{
  Internal function for simultaneous bootstrap confidence intervals
}
\description{
  Internal function for simultaneous bootstrap confidence intervals
  based on resampled residuals
}
\references{
  Westfall, P. H. and Young, S. S. (1993) Resampling-Based
    Multiple Testing: Examples and Methods for \eqn{p}--Value
    Adjustment. New York: Wiley.
}
\note{
  Only internal function. Use function \code{\link{sbdiv}} instead
}
\keyword{ htest }
